requirements.txt
albumentations==1.0.3
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
cycler==0.10.0
docopt==0.6.2
flatbuffers==2.0
googledrivedownloader==0.4
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
Jinja2==3.0.1
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.3
Nib==0.5.7
nibabel==3.2.1
numpy==1.21.2
onnx==1.10.1
onnxruntime==1.8.1
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
packaging==21.0
Pillow==8.3.2
pipreqs==0.4.10
protobuf==3.18.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.1
sh==1.14.2
six==1.16.0
tifffile==2021.8.30
torch==1.9.0
torchvision==0.10.0
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.7
yarg==0.1.9
zipp==3.5.0
